% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconvergent-fonts.R
\docType{data}
\name{original_df}
\alias{original_df}
\alias{smooth_df}
\title{Data.frame representation of strokes in Inconvergent's "original" vector font
\url{https://github.com/inconvergent/gridfont}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 438 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 696 rows and 7 columns.
}
\usage{
original_df

smooth_df
}
\description{
Data.frame representation of strokes in Inconvergent's "original" vector font
\url{https://github.com/inconvergent/gridfont}
}
\keyword{datasets}
