% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconvergent-fonts.R
\docType{data}
\name{original_json}
\alias{original_json}
\alias{smooth_json}
\title{Original JSON fonts by Inconvergent}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 5.
}
\usage{
original_json

smooth_json
}
\description{
\url{https://github.com/inconvergent/gridfont}
}
\keyword{datasets}
